+++ shardis-ui/src/main/frontend/app/services/auth.service.ts
@@ -6,13 +6,19 @@ export class AuthService {
 
   private authenticated:boolean = false;
   private tokenData:Oauth2TokenData = new Oauth2TokenData();
+  private tokenExpirationDate:Date = null;
   private userData:any = null;
 
   constructor(public http:Http) {
     if (window.localStorage.getItem('tokenData')) {
       this.tokenData = JSON.parse(window.localStorage.getItem('tokenData'));
       this.authenticated = true;
-      this.userData = this.fetchUserData();
+      this.userData = this.decodeAccessToken(this.tokenData.access_token);
+      this.tokenExpirationDate = new Date(this.userData.exp * 1000);
+      if (this.tokenExpirationDate < new Date()) {
+        console.log('Session timeout');
+        this.logout();
+      }
     }
   }
 
@@ -42,9 +48,9 @@ export class AuthService {
         data => {
           this.tokenData = data.json();
           window.localStorage.setItem('tokenData', JSON.stringify(this.tokenData));
-          console.log(this.tokenData);
           this.authenticated = true;
-          this.userData = this.fetchUserData();
+          this.userData = this.decodeAccessToken(this.tokenData.access_token);
+          this.tokenExpirationDate = new Date(this.userData.exp * 1000);
         },
         err => {
           console.log(err);
@@ -63,10 +69,20 @@ export class AuthService {
     return this.userData;
   }
 
-  private fetchUserData() {
+  public getTokenExpirationDate():Date {
+    return this.tokenExpirationDate;
+  }
+
+  public getAuthorizationHeaders():Headers {
     var authorizationHeaders = new Headers();
-    authorizationHeaders.append('Authorization', `Bearer ${this.tokenData.access_token}`);
-    this.http.get('/api/user', {headers: authorizationHeaders})
+    if (this.authenticated) {
+      authorizationHeaders.append('Authorization', `Bearer ${this.tokenData.access_token}`);
+    }
+    return authorizationHeaders;
+  }
+
+  private fetchUserData() {
+    this.http.get('/api/user', {headers: this.getAuthorizationHeaders()})
       .subscribe(
         data => {
           this.userData = data.json();
@@ -75,14 +91,18 @@ export class AuthService {
       );
   }
 
+  private decodeAccessToken(access_token:string) {
+    return JSON.parse(window.atob(access_token.split('.')[1]));
+  }
 }
 
 class Oauth2TokenData {
   access_token:string = null;
   token_type:string = null;
-  expires_in:string = null;
+  expires_in:number = null;
   scope:string = null;
   jti:string = null;
+  refresh_token:string = null;
 
   constructor() {
   }