+++ shardis-ui/src/main/frontend/app/views/home/home.component.ts
@@ -1,5 +1,4 @@
 import {Component} from 'angular2/core';
-import {FORM_DIRECTIVES} from 'angular2/common';
 
 import {Title} from './services/title.service.ts';
 import {XLarge} from './../../directives/x-large.directive.ts';
@@ -15,7 +14,7 @@ import {XLarge} from './../../directives/x-large.directive.ts';
   ],
   // We need to tell Angular's compiler which directives are in our template.
   // Doing so will allow Angular to attach our behavior to an element
-  directives: [...FORM_DIRECTIVES, XLarge as any],
+  directives: [XLarge as any],
   // We need to tell Angular's compiler which custom pipes are in our template.
   pipes: [],
   // Our list of styles in our component. We may add more to compose many styles together