+++ shardis-ui/src/main/frontend/app/views/home/services/title.service.ts
@@ -1,11 +1,12 @@
 import {Injectable} from 'angular2/core';
 import {Http} from 'angular2/http';
+import {AuthService} from '../../../services/auth.service';
 
 @Injectable()
 export class Title {
   value = 'Angular 2';
 
-  constructor(public http:Http) {
+  constructor(public http:Http, public authService:AuthService) {
   }
 
   getData() {
@@ -13,8 +14,8 @@ export class Title {
 
     return this
       .http
-      .get('/api/title')
-      .map(res =>  res.json());
+      .get('/api/title', {headers: this.authService.getAuthorizationHeaders()})
+      .map(res => res.json());
   }
 
 }